+++ test/pyinstall-aux-static-exec/auxlist.py
@@ -1,0 +1,6 @@
+aux_apps = (
+    'aux-glibc-dynamic',
+    'aux-glibc-static',
+    'aux-musl-dynamic',
+    'aux-musl-static',
+)